requirements.txt
Django==4.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
django-channels==0.7.0
django-cors-headers==3.14.0
asgiref==3.8.0
channels-redis==3.0.0
djangorestframework-gis==1.1
pyproj==3.5.0
psycopg2-binary==2.9.10
daphne==2.3